flightWebapp/requirements.txt
alembic==0.9.9
certifi==2018.4.16
click==6.7
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mkl-fft==1.0.4
mkl-random==1.0.1
numpy==1.15.0
pandas==0.23.4
psycopg2==2.7.5
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.10
Werkzeug==0.14.1
wincertstore==0.2
